playwright.rye/pyproject.toml
[project]
name = "samproj"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rhie-coder", email = "quotia72@naver.com" }
]
dependencies = [
    "pytest-playwright>=0.5.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"

[tool.rye.scripts]
thead = 'pytest --headed --slowmo 1000'

    # '--browser=chromium',
    # '--browser=firefox',
    # '--browser=webkit',

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/samproj"]
